Clang reported use of sqrt as ambigious

+++ Simulation/FluidSimulation.cpp
@@ -407,10 +407,10 @@ void FluidSimulation::simulateErosion(double dt)
             sinAlpha = std::max(sinAlpha,0.1f);
 
             // local sediment capacity of the flow
-            float capacity = Kc * sqrt(uV*uV+vV*vV)*sinAlpha*(std::min(water(y,x),0.01f)/0.01f) ;
+            float capacity = Kc * sqrtf(uV*uV+vV*vV)*sinAlpha*(std::min(water(y,x),0.01f)/0.01f) ;
             float delta = (capacity-sediment(y,x));
 
-            float v = sqrt(uV*uV+vV*vV);
+            float v = sqrtf(uV*uV+vV*vV);
             float fctr = (std::min(water(y,x),0.01f)/0.01f);
 
             if (delta > 0.0f)
